Add support for Bungee portals with coordinates
BungeeCord exposes a special Forward channel for plugins to send
messages to plugins on other servers. This channel allows to perform a
cross-server teleport that includes a location. This teleport is
performed in two steps: 1) the user is connected to the new server 2) a
message is sent to the new server requesting that the player be
teleported to the required location (this is done by serializing the
portal across the wire and performing a "local" teleport with
that portal).

To perform this generalization, the distinction between Local and Server
portals has been removed (as now a Server portal supports all the
features of a Local portal). The distinction between teleport event
types has been kept.

Additionally, to prevent dangerous portals when crossing servers, the
teleport location on the server being left has been changed to a safe
spot nearby the triggered teleport (it moves the player two blocks
backwards). This keeps the behavior reasonable in the case where the
connection to the new server fails (or if the old server is directly
rejoined).