Use nearby for Bungee portal previous location
To prevent dangerous portals when crossing servers, the
teleport location on the server being left has been changed to a safe
spot nearby the triggered teleport (it moves the player two blocks
backwards) instead of the world spawn point. This keeps the user in
roughly the same place in case the connection to the new server fails
(or if the old server is directly rejoined).